% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_records.R
\name{add_diag_data}
\alias{add_diag_data}
\title{#' @title Format Diagnostic Records
 @description \code{format_diag()} formats diagnostic records by extracting relevant columns and renaming them.
 @param diag_data A data frame containing diagnostic records.
 @param match_status A data frame containing the mapping of smoking status.
 @param id A string indicating the column name for patient ID (default is "mrn").
 @param note_date A string indicating the column name for note date (default is "ct_scan_date").
 @param smoking_status A string indicating the column name for smoking status (default is "smoking").
 @param quit_date A string indicating the column name for quit date (default is NULL).
 @param pack_days A string indicating the column name for pack days (default is NULL).
 @param pack_years A string indicating the column name for pack years (default is NULL).
 @return A data frame with formatted diagnostic records containing columns for patient ID, note date, smoking status, quit date, pack days, and pack years.}
\usage{
add_diag_data(smoking_data, diag_data, ...)
}
\description{
#' @title Format Diagnostic Records
 @description \code{format_diag()} formats diagnostic records by extracting relevant columns and renaming them.
 @param diag_data A data frame containing diagnostic records.
 @param match_status A data frame containing the mapping of smoking status.
 @param id A string indicating the column name for patient ID (default is "mrn").
 @param note_date A string indicating the column name for note date (default is "ct_scan_date").
 @param smoking_status A string indicating the column name for smoking status (default is "smoking").
 @param quit_date A string indicating the column name for quit date (default is NULL).
 @param pack_days A string indicating the column name for pack days (default is NULL).
 @param pack_years A string indicating the column name for pack years (default is NULL).
 @return A data frame with formatted diagnostic records containing columns for patient ID, note date, smoking status, quit date, pack days, and pack years.
}
