% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\docType{data}
\name{color}
\alias{color}
\title{Scatter Plot of Smoking Data}
\format{
An object of class \code{character} of length 4.
}
\usage{
color
}
\arguments{
\item{data_list}{A list containing data frames for each patient, where each data frame includes columns for note date,
the variable of interest (e.g., pack years), and smoking status.}

\item{patient}{A string specifying the patient identifier for whom the plot is to be generated.}

\item{variable}{A string specifying the variable to be plotted (default is "pack_years").}

\item{smoking_status}{A string specifying the column name for smoking status (default is "smoking_status").}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
Creates a scatter plot of smoking data for a specific patient, showing the variable (e.g., pack years) over time, colored by smoking status.
}
\keyword{datasets}
