% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_status}
\alias{heatmap_status}
\title{Heatmap of Smoking Status Over Time}
\usage{
heatmap_status(data_list, patients, ...)
}
\arguments{
\item{data_list}{A list containing data frames for each patient, where each data frame includes columns for note date and smoking status.}

\item{patients}{A vector of patient identifiers for whom the heatmap is to be generated.}

\item{...}{Additional arguments passed to the Heatmap function from the ComplexHeatmap package.}
}
\value{
A heatmap object representing the smoking status of patients over time.
}
\description{
Creates a heatmap of smoking status for multiple patients over time.
}
