% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_packs.R
\name{longitudinal_packs}
\alias{longitudinal_packs}
\title{Longitudinal-level quantitative Correction}
\usage{
longitudinal_packs(
  data_list,
  cluster_col = "inconsistent",
  year_cap = 20,
  day_cap = 200,
  per_years = 10,
  per_years_current = 2
)
}
\arguments{
\item{data_list}{a list of data frames contains patient's smoking data.}

\item{cluster_col}{a character string indicating the column name used to identify clusters of inconsistent records,
default is "inconsistent".}

\item{year_cap}{a numeric value indicating the maximum cap for pack-years, default is 20.}

\item{day_cap}{a numeric value indicating the maximum cap for pack-days, default is 200.}

\item{per_years}{a numeric value indicating the threshold for pack-years increase per year for current smokers, default is 10.}

\item{per_years_current}{a numeric value indicating the threshold for pack-years increase per year for non-current smokers, default is 2.}
}
\value{
A list of corrected data frames for each patient.
}
\description{
\code{longitudinal_packs()} Corrects the longitudinal quantitative variables
             (e.g. pack-years or pack-days) for each patient.
}
