% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_smoking.R
\name{longitudinal_smoking}
\alias{longitudinal_smoking}
\title{Longitudinal Smoking Status Correction}
\usage{
longitudinal_smoking(data_list, percent = 0.25)
}
\arguments{
\item{data_list}{a list of data frames contains patient's smoking data.}

\item{percent}{a numeric value indicating the percentage of non-zero records quantitative variables (e.g. pack-years or pack-days)
to determine whether to classify a patient as "Never" or "Former" smoker, default is 0.25.}
}
\value{
A list of corrected data frames for each patient.
}
\description{
\code{longitudinal_smoking()} Corrects the longitudinal smoking status for each patient.
}
