% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process Smoking Data}
\usage{
preprocess(
  data,
  match_status,
  id = "mrn",
  note_date = "noted_date",
  quit_date = "smoking_quit_date",
  smoking_status = c("smoking_tob_use", "tobacco_user"),
  merge_status = "smoking_status",
  pack_days = "TOBACCO_PAK_PER_DY",
  pack_years = "tobacco_pck_yrs_old",
  time_format = "\%m/\%d/\%y",
  other = NA
)
}
\arguments{
\item{data}{a data frame containing smoking data for a single patient.}

\item{match_status}{a data frame containing the mapping of smoking status.}

\item{id}{a string indicating the column name for patient ID, default is "mrn".}

\item{note_date}{a string indicating the column name for note date, default is "noted_date".}

\item{quit_date}{a string indicating the column name for quit date, default is "smoking_quit_date".}

\item{smoking_status}{a character vector indicating the column names for smoking status, default is c("smoking_tob_use", "tobacco_user").}

\item{merge_status}{a string indicating the column name for merged smoking status, default is "smoking_status".}

\item{pack_days}{a string indicating the column name for pack days, default is "TOBACCO_PAK_PER_DY".}

\item{pack_years}{a string indicating the column name for pack years, default is "tobacco_pck_yrs_old".}

\item{other}{a string indicating any other column names to be included, default is NA.}
}
\value{
A list of corrected data frames for each patient.
}
\description{
\code{preprocess()} Pre-processes smoking data for a single patient.
}
