% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_smoking.R
\name{record_smoking}
\alias{record_smoking}
\title{Record-Level Smoking Status}
\usage{
record_smoking(
  data_list,
  quit_date = "quit_date",
  note_date = "note_date",
  percent = 0.1
)
}
\arguments{
\item{data_list}{a list of data frames contains patient's smoking data.}

\item{quit_date}{a string indicating the column name for quit date, default is "quit_date".}

\item{note_date}{a string indicating the column name for note date, default is "note_date".}

\item{percent}{a numeric value indicating the percentage of non-zero records quantitative variables (e.g. pack-years or pack-days)
to determine whether to classify a patient as "Never" or "Former" smoker, default is 0.1.}
}
\value{
A list of corrected data frames for each patient.
}
\description{
\code{record_smoking()} Identifies the record-level inconsistency in smoking status.
}
